Fix bug in Device::Finalize: Forgot to clear streams (#3143)
## Summary

We forgot to clear the non-owning gpu_stream vector. They still held the
old streams after we destroyed the streams in Device::Finalize. This bug
affects codes that call amrex::Initialize and amrex::Finalize
repeatedly.

Note that we have the non-owning gpu stream vector for OMP CPU thread
safety.

The bug was introduced in #1064.

## Additional background

#3141 

## Checklist

The proposed changes:
- [x] fix a bug or incorrect behavior in AMReX
- [ ] add new capabilities to AMReX
- [ ] changes answers in the test suite to more than roundoff level
- [ ] are likely to significantly affect the results of downstream AMReX
users
- [ ] include documentation in the code and/or rst files, if appropriate