amrex::Math: No longer need to use sycl math functions (#3164)
We introduced `amrex::Math::floor`, etc. because their std versions were
not supported in device code by the Intel compiler at that time.
Otherwise, the user would have to call either sycl::floor or std::floor
with the help of macros. This is no longer the case. Now we can switch
to use the std versions. The amrex::Math versions are kept for backward
compatibility. The particle code is not updated because it's undergoing
some large changes.