Run PODVector::push_back on GPU (#3308)
There are two versions of PODVector::push_back. The `const&` version
previously ran on CPU, whereas the `&&` version on GPU. It seems that
they should have the same behavior. Thus the `const&` version now runs
on GPU.

(We may not even need to have two versions, because the move (i.e.,
`&&`) version does not move. It calls another function that takes
`const&`.)