Fix a memory "leak" in VisMF's persistent streams (#3592)
## Summary

VisMF uses a std::map to keep track of the streams it opens. Each
PersistentIFStream object also has a vector used as IO buffer. When a
stream is closed, we need to free to memory used by the buffer. The
issue was it used std::vector::clear(), which only changes size()
without reducing capacity() at all. So the memory was never freed. The
issue is now fixed by std::vector::swap.

## Additional background

Thank Simon Guichandut (@simonguichandut) for reporting this! Thank Eric
Johnson (@yut23) for pinning down the location of the leak with the help
of the Massif heap profiler!

## Checklist

The proposed changes:
- [x] fix a bug or incorrect behavior in AMReX
- [ ] add new capabilities to AMReX
- [ ] changes answers in the test suite to more than roundoff level
- [ ] are likely to significantly affect the results of downstream AMReX
users
- [ ] include documentation in the code and/or rst files, if appropriate